refactor(user-management): relocate role.py to autobot_shared, cosmetic diff only (#12647)
